+++ speeduino/board_avr2560.ino
@@ -117,9 +117,9 @@ ISR(TIMER5_OVF_vect)
     ++timer5_overflow_count;
 }
 
-static inline unsigned long micros_safe()
+static inline uint32_t micros_safe()
 {
-  unsigned long newMicros;
+  uint32_t newMicros;
   noInterrupts();
   newMicros = (((timer5_overflow_count << 16) + TCNT5) * 4);
   interrupts();
