+++ speeduino/corrections.h
@@ -45,7 +45,7 @@ extern uint8_t activateMAPDOT; //The mapDOT value seen when the MAE was activate
 extern uint8_t activateTPSDOT; //The tpsDOT value seen when the MAE was activated. 
 
 extern uint16_t AFRnextCycle;
-extern unsigned long knockStartTime;
+extern uint32_t knockStartTime;
 extern uint8_t lastKnockCount;
 extern int16_t knockWindowMin; //The current minimum crank angle for a knock pulse to be valid
 extern int16_t knockWindowMax;//The current maximum crank angle for a knock pulse to be valid
